requirements.txt
streamlit
requests
Pillow
python-dotenv